+++ Frends.Salesforce.ExecuteQuery/Frends.Salesforce.ExecuteQuery/Definitions/AuthenticationMethod.cs
@@ -10,11 +10,11 @@ public enum AuthenticationMethod
     /// </summary>
     AccessToken = 0,
     /// <summary>
-    /// Authenticate by providing required informations to fetch OAuth2 access token.
+    /// Authenticate by providing required information to fetch OAuth2 access token.
     /// </summary>
     OAuth2WithPassword = 1,
     /// <summary>
-    /// Authenticate by providing required informations to fetch OAuth2 access token client_credentials.
+    /// Authenticate by providing required information to fetch OAuth2 access token client_credentials.
     /// </summary>
     OAuth2WithClientCredentials = 2
 }