+++ Frends.Salesforce.ExecuteQuery/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## [2.3.0] - 2026-01-31
-### Changed
-- Added new authentication method for authentication with client credentials
+### Added
+- Added new authentication method for authentication by OAuth2 with client credentials
 
 ## [2.2.0] - 2025-12-12
 ### Changed